extract `InteractEvent` and `Interaction` to modules
